+++ api/tests/team/integration/infrastructure/repositories/organization-invitation.repository.test.js
@@ -1,12 +1,12 @@
 import _ from 'lodash';
 
-import { NotFoundError } from '../../../../lib/domain/errors.js';
-import { Membership } from '../../../../lib/domain/models/Membership.js';
-import { OrganizationInvitation } from '../../../../lib/domain/models/OrganizationInvitation.js';
-import * as organizationInvitationRepository from '../../../../lib/infrastructure/repositories/organization-invitation-repository.js';
-import { catchErr, databaseBuilder, expect, knex, sinon } from '../../../test-helper.js';
+import { NotFoundError } from '../../../../../lib/domain/errors.js';
+import { Membership } from '../../../../../lib/domain/models/Membership.js';
+import { OrganizationInvitation } from '../../../../../lib/domain/models/OrganizationInvitation.js';
+import { organizationInvitationRepository } from '../../../../../src/team/infrastructure/repositories/organization-invitation.repository.js';
+import { catchErr, databaseBuilder, expect, knex, sinon } from '../../../../test-helper.js';
 
-describe('Integration | Repository | OrganizationInvitationRepository', function () {
+describe('Integration | Team | Infrastructure | Repository | organization-invitation', function () {
   describe('#create', function () {
     let organizationId;
